Fixed bugs and added new features.

+++ applications/routers/delete_analysis.py
@@ -1,24 +1,30 @@
-from flask import redirect, url_for, flash
+from flask import redirect, url_for, flash, request
 from flask_login import current_user
 
 #-------------User Packages --------------------#
 from applications import app
-from applications.models import Research
-from applications.database import db
 
-
-
-@app.route('/delete_analysis/<id>', methods=["GET","POST"])
+@app.route('/delete_analysis/<id>', methods=["GET", "POST"])
 def delete_analysis(id):
 
-    try:
-        db.session.rollback()
-        db.session.begin()
-        Research.query.filter(Research.user_id == current_user.id, Research.id == id).delete()
-        db.session.commit()
-    except Exception as e:
-        flash(f"Something went wrong while deleting from the data to the database. error: {e}", category='warning')
+    if request.method == "POST":
+    
+        print()
+        print("inside the delete research")
+        print()
         return redirect(url_for('analytics_page'))
 
-    flash("Successfully deleted!", category='success')
-    return redirect(url_for('analytics_page'))
+    #-------------User Packages --------------------#
+        from applications.models import Research
+        from applications.database import db
+        try:
+            Research.query.filter(Research.user_id == current_user.id, Research.id == id).delete()
+            db.session.commit()
+            flash("Successfully deleted!", category='success')
+            return redirect(url_for('analytics_page'))
+
+        except Exception as e:
+            flash(f"Something went wrong while deleting from the data to the database. error: {e}", category='warning')
+            return redirect(url_for('analytics_page'))
+    else:
+        return redirect(url_for('analytics_page'))
